SCMOD-7821: Converters refactored
