+++ caf-logging-logback-access-converters/pom.xml
@@ -27,14 +27,16 @@
         <version>1.2.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>caf-logging-logback</artifactId>
+    <artifactId>caf-logging-logback-access-converters</artifactId>
 
     <dependencies>
         <dependency>
+            <groupId>ch.qos.logback</groupId>
+            <artifactId>logback-access</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.github.cafapi.logging</groupId>
-            <artifactId>caf-logging-logback-converters</artifactId>
-            <scope>runtime</scope>
+            <artifactId>caf-logging-common</artifactId>
         </dependency>
     </dependencies>
-
 </project>
