+++ caf-logging-logback/src/main/resources/logback.xml
@@ -17,11 +17,11 @@
 -->
 <configuration>
     <conversionRule conversionWord="aProcessId"
-                    converterClass="com.github.cafapi.logging.logback.ProcessAndThreadIdPatternConverter" />
+                    converterClass="com.github.cafapi.logging.logback.converters.ProcessAndThreadIdPatternConverter" />
     <conversionRule conversionWord="sanitize"
-                    converterClass="com.github.cafapi.logging.logback.SanitizeConverter" />
+                    converterClass="com.github.cafapi.logging.logback.converters.SanitizeConverter" />
     <conversionRule conversionWord="maybeJsonMsgAndEx"
-                    converterClass="com.github.cafapi.logging.logback.MaybeJsonMessageConverter" />
+                    converterClass="com.github.cafapi.logging.logback.converters.MaybeJsonMessageConverter" />
 
     <include optional="true" resource="includedLogback.xml" />
