+++ pom.xml
@@ -67,7 +67,8 @@
         <module>caf-logging-common</module>
         <module>caf-logging-log4j2</module>
         <module>caf-logging-logback</module>
-        <module>caf-logging-logback-access</module>
+        <module>caf-logging-logback-converters</module>
+        <module>caf-logging-logback-access-converters</module>
     </modules>
 
     <properties>
@@ -88,6 +89,11 @@
             <dependency>
                 <groupId>com.github.cafapi.logging</groupId>
                 <artifactId>caf-logging-common</artifactId>
+                <version>1.2.0-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>com.github.cafapi.logging</groupId>
+                <artifactId>caf-logging-logback-converters</artifactId>
                 <version>1.2.0-SNAPSHOT</version>
             </dependency>
             <dependency>
